backend/requirements.txt

fastapi
uvicorn
tortoise-orm
aiosqlite
pydantic-ai
python-dotenv
pyjwt
passlib[bcrypt]
python-multipart
pydantic_settings